Clean up sphinx conf file
