+++ docs/source/conf/10-project.py
@@ -1,0 +1,14 @@
+from scico._version import package_version
+
+# General information about the project.
+project = "SCICO"
+copyright = "2020-2023, SCICO Developers"
+
+# The version info for the project you're documenting, acts as replacement for
+# |version| and |release|, also used in various other places throughout the
+# built documents.
+#
+# The short X.Y version.
+version = package_version()
+# The full version, including alpha/beta/rc tags.
+release = version
