+++ docs/source/conf/20-extensions.py
@@ -1,0 +1,19 @@
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+extensions = [
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autodoc",
+    "sphinx_autodoc_typehints",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.doctest",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode",
+    "sphinxcontrib.bibtex",
+    "sphinx.ext.inheritance_diagram",
+    "sphinx.ext.todo",
+    "nbsphinx",
+]
+
+
+bibtex_bibfiles = ["references.bib"]
