+++ docs/source/conf/40-intersphinx.py
@@ -1,0 +1,13 @@
+# Intersphinx mapping
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3/", None),
+    "numpy": ("https://numpy.org/doc/stable/", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/", None),
+    "matplotlib": ("https://matplotlib.org/stable/", None),
+    "jax": ("https://jax.readthedocs.io/en/latest/", None),
+    "flax": ("https://flax.readthedocs.io/en/latest/", None),
+    "ray": ("https://docs.ray.io/en/latest/", None),
+    "svmbir": ("https://svmbir.readthedocs.io/en/latest/", None),
+}
+# Added timeout due to periodic scipy.org down time
+# intersphinx_timeout = 30
