+++ docs/source/conf/50-graphviz.py
@@ -1,0 +1,15 @@
+graphviz_output_format = "svg"
+inheritance_graph_attrs = dict(rankdir="LR", fontsize=9, ratio="compress", bgcolor="transparent")
+inheritance_edge_attrs = dict(
+    color='"#2962ffff"',
+)
+inheritance_node_attrs = dict(
+    shape="box",
+    fontsize=9,
+    height=0.4,
+    margin='"0.08, 0.03"',
+    style='"rounded,filled"',
+    color='"#2962ffff"',
+    fontcolor='"#2962ffff"',
+    fillcolor='"#f0f0f8b0"',
+)
