+++ docs/source/conf/55-nbsphinx.py
@@ -1,0 +1,17 @@
+nbsphinx_execute = "never"
+nbsphinx_prolog = """
+.. raw:: html
+
+    <style>
+    .nbinput .prompt, .nboutput .prompt {
+        display: none;
+    }
+    div.highlight {
+        background-color: #f9f9f4;
+    }
+    p {
+        margin-bottom: 0.8em;
+        margin-top: 0.8em;
+    }
+    </style>
+"""
