+++ docs/source/conf/71-texinfo.py
@@ -1,0 +1,16 @@
+# -- Options for Texinfo output -------------------------------------------
+
+# Grouping the document tree into Texinfo files. List of tuples
+# (source start file, target name, title, author,
+#  dir menu entry, description, category)
+texinfo_documents = [
+    (
+        "index",
+        "SCICO",
+        "SCICO Documentation",
+        "SCICO Developers",
+        "SCICO",
+        "Scientific Computational Imaging COde (SCICO)",
+        "Miscellaneous",
+    ),
+]
