+++ docs/source/conf/72-man_page.py
@@ -1,0 +1,8 @@
+# -- Options for manual page output ---------------------------------------
+
+# One entry per manual page. List of tuples
+# (source start file, name, description, authors, manual section).
+man_pages = [("index", "scico", "SCICO Documentation", ["SCICO Developers"], 1)]
+
+# If true, show URL addresses after external links.
+# man_show_urls = False
